Move apps out of the launcher package
