+++ workshops/InversionOfControl/default/app/src/main/java/com/novoda/iockata/apps/AppAdapter.kt
@@ -1,4 +1,4 @@
-package com.novoda.iockata.launcher.apps
+package com.novoda.iockata.apps
 
 import android.app.Activity
 import android.support.v7.widget.RecyclerView
@@ -7,10 +7,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import com.novoda.iockata.R
 import com.novoda.iockata.core.devicefeatures.DeviceFeature
-import com.novoda.iockata.launcher.apps.browser.BrowserApp
-import com.novoda.iockata.launcher.apps.camera.CameraApp
-import com.novoda.iockata.launcher.apps.phone.PhoneApp
-import com.novoda.iockata.launcher.apps.toaster.ToasterApp
+import com.novoda.iockata.apps.browser.BrowserApp
+import com.novoda.iockata.apps.camera.CameraApp
+import com.novoda.iockata.apps.phone.PhoneApp
+import com.novoda.iockata.apps.toaster.ToasterApp
 
 class AppAdapter(activity: Activity, deviceFeature: DeviceFeature) : RecyclerView.Adapter<AppAdapter.ViewHolder>() {
