+++ workshops/InversionOfControl/default/app/src/main/java/com/novoda/iockata/apps/toaster/ToasterApp.kt
@@ -1,9 +1,9 @@
-package com.novoda.iockata.launcher.apps.toaster
+package com.novoda.iockata.apps.toaster
 
 import android.app.Activity
 import android.widget.Toast
 import com.novoda.iockata.R
-import com.novoda.iockata.launcher.apps.App
+import com.novoda.iockata.apps.App
 
 class ToasterApp(private val activity: Activity) : App {
